+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -238,7 +238,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   bytes: z.number(),
   file_name: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2024-11-24T21:26:36.345Z",
+    "2024-11-25T07:27:21.148Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -269,7 +269,7 @@ export const FileUploadResponseBody$outboundSchema: z.ZodType<
   purpose: FileUploadPurpose$outboundSchema,
   bytes: z.number(),
   fileName: z.string(),
-  created: z.date().default(() => new Date("2024-11-24T21:26:36.345Z"))
+  created: z.date().default(() => new Date("2024-11-25T07:27:21.148Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {