+++ packages/orq-rc/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v1.42.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v1.42.0-rc.3] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.42.0-rc.3 - packages/orq-rc
+
+## 2024-11-25 07:38:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.444.4 (2.461.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.42.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v1.42.0-rc.5] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.42.0-rc.5 - packages/orq-rc